Restructure app, add DB service and Docker changes
Move project source into app/, replace route modules with express routers under app/models, and centralize DB access in a new app/services/db.js (mysql2 + dotenv). Update app entry (app/app.js) to use the new service paths and serve static files from /static. Add several Pug views (users, user, pubs, pub) and update layout to reference /static/styles.css. Add index.js entrypoint. Revise Dockerfile to use /src, install supervisor, and copy project files; update docker-compose.yml to use env_file, supervisor command, configurable ports, and mount schema.sql for DB initialization. Update .gitignore/.dockerignore and package metadata/dependencies accordingly. Remove old src/config, removed DB schema/seed from database/ and added schema.sql at repo root.
