+++ .dockerignore
@@ -1,6 +1,3 @@
+.git
 node_modules
 npm-debug.log
-.git
-.gitignore
-Dockerfile
-README.md
