+++ app/views/index.pug
@@ -1,4 +1,4 @@
-extends ../layout/default
+extends layout
 
 block content
