+++ app/views/layout.pug
@@ -2,7 +2,7 @@ doctype html
 html
   head
     title= title
-    link(rel="stylesheet", href="/css/styles.css")
+    link(rel="stylesheet", href="/static/styles.css")
   body
     header
       a(href="/")
