+++ app/views/user.pug
@@ -1,0 +1,20 @@
+extends layout
+
+block content
+  main.container
+    h1= user.first_name + ' ' + user.last_name
+    p Email: #{user.email || 'No email'}
+    p Joined: #{user.created_at ? new Date(user.created_at).toLocaleDateString() : 'Unknown'}
+
+    h2 Reviews  
+    if reviews.length
+      each review in reviews
+        section
+          p Pub: #{review.pub_name}
+          p Beer: #{review.beer_name}
+          p Rating: #{review.rating}
+          if review.ai_pour_score
+            p AI Pour Score: #{review.ai_pour_score}
+          p= review.comment
+    else
+      p No reviews yet.
