+++ docker-compose.yml
@@ -1,48 +1,34 @@
+version: '3.3'
 services:
-  app:
-    build: .
-    container_name: tapthat_app
-    ports:
-      - "3000:3000"
+  web:
+    build:
+      context: ./
     volumes:
-      - .:/usr/src/app
-      - /usr/src/app/node_modules
-    environment:
-      - DB_HOST=db
-      - DB_USER=tapuser
-      - DB_PASSWORD=tappass
-      - DB_NAME=tapthat
+      - .:/src
+      - /src/node_modules
+    command: supervisor -w index.js,app,static -e js,pug,html,css index.js
+    ports:
+      - "${WEB_PORT:-3000}:3000"
     depends_on:
       - db
-
   db:
-    image: mysql:8
-    container_name: tapthat_db
+    image: mysql
     restart: always
-    environment:
-      MYSQL_ROOT_PASSWORD: example
-      MYSQL_DATABASE: tapthat
-      MYSQL_USER: tapuser
-      MYSQL_PASSWORD: tappass
+    env_file:
+      - ./.env
     ports:
-      - "3306:3306"
+      - "${DB_PUBLISHED_PORT:-3308}:3306"
     volumes:
       - db_data:/var/lib/mysql
-      - ./database/schema.sql:/docker-entrypoint-initdb.d/1.sql
-      - ./database/seed.sql:/docker-entrypoint-initdb.d/2.sql
-
+      - ./schema.sql:/docker-entrypoint-initdb.d/001-schema.sql:ro
   phpmyadmin:
-    image: phpmyadmin/phpmyadmin
-    container_name: tapthat_phpmyadmin
+    image: phpmyadmin/phpmyadmin:latest
     restart: always
-    ports:
-      - "8080:80"
     environment:
       PMA_HOST: db
       PMA_PORT: 3306
-      MYSQL_ROOT_PASSWORD: example
-    depends_on:
-      - db
-
+      MYSQL_ROOT_PASSWORD: ${MYSQL_ROOT_PASSWORD}
+    ports:
+      - "${PHPMYADMIN_PORT:-8081}:80"
 volumes:
   db_data:
