+++ package.json
@@ -1,30 +1,20 @@
 {
-  "name": "softwareengineeringmodulecoursework",
+  "name": "sd2-example",
   "version": "1.0.0",
-  "description": "Full stack community web application built with Node.js, Express, MySQL, and PUG. The platform enables users to share goods or services cooperatively for mutual benefit. Developed using Scrum, Docker, GitHub Projects, and CI CD with GitHub Actions, focusing on usability, security, and ethical design.",
+  "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "nodemon src/app.js",
-    "start": "node src/app.js"
+    "start": "supervisor index.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jackoturner/SoftwareEngineeringModuleCoursework.git"
-  },
-  "keywords": [],
   "author": "",
   "license": "ISC",
-  "type": "commonjs",
-  "bugs": {
-    "url": "https://github.com/jackoturner/SoftwareEngineeringModuleCoursework/issues"
-  },
-  "homepage": "https://github.com/jackoturner/SoftwareEngineeringModuleCoursework#readme",
   "dependencies": {
-    "express": "^5.2.1",
-    "mysql2": "^3.15.0",
-    "pug": "^3.0.3"
-  },
-  "devDependencies": {
-    "nodemon": "^3.1.11"
+    "dotenv": "^10.0.0",
+    "express": "^4.17.1",
+    "mysql": "^2.18.1",
+    "mysql2": "^2.3.0",
+    "pug": "^3.0.3",
+    "supervisor": "^0.12.0"
   }
 }
