--- src/config/db.js
@@ -1,15 +1,0 @@
-const mysql = require("mysql2");
-
-const pool = mysql.createPool({
-  host: process.env.DB_HOST || "db",
-  user: process.env.DB_USER || "root",
-  password: process.env.DB_PASSWORD || "example",
-  database: process.env.DB_NAME || "tapthat",
-  waitForConnections: true,
-  connectionLimit: 10,
-  queueLimit: 0
-});
-
-console.log("MySQL pool created");
-
-module.exports = pool;
