--- src/views/pub.pug
@@ -1,13 +1,0 @@
-extends ../layout/default
-
-block content
-  main.container
-    h1= pub.name
-    p Location: #{pub.location}
-    h2 Reviews
-    each review in reviews
-      section
-        p User: #{review.username}
-        p Beer: #{review.beer_name}
-        p Rating: #{review.rating}
-        p= review.review_text
