--- src/views/pubs.pug
@@ -1,9 +1,0 @@
-extends ../layout/default
-
-block content
-    main.container
-        h1 Pubs
-        p Temporary page for testing purposes.
-        each pub in pubs
-            section
-                a(href=`/pubs/${pub.id}`)= pub.name
