--- src/views/user.pug
@@ -1,19 +1,0 @@
-extends ../layout/default
-
-block content
-  main.container
-    h1= user.username
-    p Email: #{user.email || 'No email'}
-    p Joined: #{user.created_at ? new Date(user.created_at).toLocaleDateString() : 'Unknown'}
-
-    h2 Reviews  
-    if reviews.length
-      ul
-        each review in reviews
-          section
-            p Pub: #{review.pub_name}
-            p Beer: #{review.beer_name}
-            p Rating: #{review.rating}
-            p= review.review_text 
-    else
-      p No reviews yet.
