--- src/views/users.pug
@@ -1,12 +1,0 @@
-extends ../layout/default
-
-block content
-  main.container
-    h1 Users
-    p Temporary page for testing purposes.
-    section
-      ul
-      each user in users
-        li
-          a(href=`/users/${user.id}`)= user.username
-        | (#{user.email || 'No email'})
